Basic improvements

+++ src/main/java/me/nonetaken/ghostblocklib/util/BlockHardness.java
@@ -236,12 +236,12 @@ public class BlockHardness {
         if (isBestTool(block, itemStack.getType())) {
             speedMultiplier = getToolSpeed(itemStack.getType());
             if (itemStack.getEnchantmentLevel(Enchantment.DIG_SPEED) > 0) {
-                speedMultiplier += Math.pow(itemStack.getEnchantmentLevel(Enchantment.DIG_SPEED), 2) + 1;
+                speedMultiplier += (float) (Math.pow(itemStack.getEnchantmentLevel(Enchantment.DIG_SPEED), 2) + 1);
             }
         }
         for (PotionEffect potionEffect : player.getActivePotionEffects()) {
             if (potionEffect.getType().equals(PotionEffectType.FAST_DIGGING)) {
-                speedMultiplier *= 0.2 * (potionEffect.getAmplifier() + 1) + 1;
+                speedMultiplier *= (float) (0.2 * (potionEffect.getAmplifier() + 1) + 1);
                 break;
             }
         }
